producer/requirements.txt
kafka-python==2.0.2
faker==30.8.2
fhir.resources==8.0.0